refactor: フォーマッタを format.ts へ統合（リファクタ計画 PR-2）
- formatDateWithEra を lib/utils.ts → lib/format.ts へ移動（挙動不変）し、
  collective-burial/DetailView.tsx の import を張り替え
- lib/utils.ts の formatDate（Intl 実装・本番未参照）と
  formatPrice（deprecated な formatCurrency ラッパ・本番未参照）を削除
- plot-detail-view.tsx のローカル formatPrice ラッパを formatCurrency 直呼びに置換
- テストを移設: formatDateWithEra → format.test.ts、削除分のテストを utils.test.ts から除去
